+++ .rustfmt.toml
@@ -1,2 +1,2 @@
-format_code_in_doc_comments = true
-wrap_comments = true
+#format_code_in_doc_comments = true
+#wrap_comments = true
