+++ Cargo.toml
@@ -17,5 +17,4 @@ tokio = { version = "1", features = ["rt", "macros", "sync"] }
 uuid = { version = "1", features = ["v4"] }
 
 [dev-dependencies]
-tokio = { version = "1", features = ["rt-multi-thread"] }
-tokio-test = "0.4"
+tokio = { version = "1", features = ["rt-multi-thread", "time"] }
