+++ src/pipeline/builder.rs
@@ -1,112 +1,14 @@
-use std::any::{type_name, Any};
 use std::collections::{HashMap, HashSet};
 use std::future::Future;
-use std::panic;
-use std::pin::Pin;
 use std::sync::Arc;
 
 use tokio::select;
-use tokio::sync::mpsc::{channel, Receiver, Sender};
-use tokio::sync::Mutex;
-use tokio::task::{yield_now, JoinError, JoinSet};
+use tokio::sync::mpsc::{channel, Receiver};
+use tokio::task::JoinSet;
 
-use crate::io::{PipeReader, PipeWriter};
-use crate::sync::Synchronizer;
-use crate::StageWorkerSignal;
-
-/// A Box that can hold any value that is [Send].
-///
-/// Values sent through pipes are trait objects of this type.
-///
-/// This type is publicly exposed as it's needed when building a pipeline stage with multiple
-/// outputs. Since each output could have a different type, it's more feasible to define the
-/// outputs to use dynamic dispatching rather that static dispatching.
-///
-/// # Examples
-///
-/// Here's an example of a closure representing the task function given to the pipeline builder
-/// when creating a "branching" stage. Three outputs are returned, each of a different type.
-/// ```
-/// use async_pipes::BoxedAnySend;
-///
-/// tokio_test::block_on(async {
-///     let task = |value: String| async move {
-///         let length: usize = value.len();
-///         let excited: String = format!("{}!", value);
-///         let odd_length: bool = length % 2 == 1;
-///
-///         Option::<Vec<Option<BoxedAnySend>>>::Some(vec![
-///             Some(Box::new(length)),
-///             Some(Box::new(excited)),
-///             Some(Box::new(odd_length)),
-///         ])
-///     };
-///
-///     // E.g.:
-///     // ...
-///     // .with_branching_stage("pipe_in", vec!["pipe_len", "pipe_excited", "pipe_odd"], <task>)
-///     // ...
-///
-///     let mut results = task("hello".to_string()).await.unwrap();
-///
-///     let length = results.remove(0).unwrap().downcast::<usize>().unwrap();
-///     let excited = results.remove(0).unwrap().downcast::<String>().unwrap();
-///     let odd_length = results.remove(0).unwrap().downcast::<bool>().unwrap();
-///
-///     assert_eq!(*length, 5usize);
-///     assert_eq!(*excited, "hello!".to_string());
-///     assert_eq!(*odd_length, true);
-/// });
-/// ```
-pub type BoxedAnySend = Box<dyn Any + Send + 'static>;
-
-type ProducerFn = Box<dyn FnMut() -> TaskFuture + Send + 'static>;
-type TaskFn = Box<dyn Fn(BoxedAnySend) -> TaskFuture + Send + Sync + 'static>;
-type IterCastFn = Box<dyn Fn(BoxedAnySend) -> Vec<BoxedAnySend> + Send + Sync + 'static>;
-
-type TaskFuture = Pin<Box<dyn Future<Output = Option<Vec<Option<BoxedAnySend>>>> + Send + 'static>>;
-
-struct ProducerStage {
-    function: ProducerFn,
-    pipes: ProducerPipeNames,
-}
-
-struct ProducerPipeNames {
-    writers: Vec<String>,
-}
-
-struct TaskStage {
-    function: TaskFn,
-    pipes: TaskPipeNames,
-}
-
-struct TaskPipeNames {
-    reader: String,
-    writers: Vec<String>,
-}
-
-struct IterStage {
-    stage_type: IterStageType,
-    caster: IterCastFn,
-    pipes: TaskPipeNames,
-}
-
-enum IterStageType {
-    Flatten,
-}
-
-struct Pipe<T> {
-    writer: PipeWriter<T>,
-    /// Use an option here to "take" it when a reader is used.
-    /// Only allow one reader per pipe.
-    reader: Option<PipeReader<T>>,
-}
-
-enum Stage {
-    Producer(ProducerStage),
-    Regular(TaskStage),
-    Iterator(IterStage),
-}
+use crate::pipeline::io::*;
+use crate::pipeline::sync::*;
+use crate::pipeline::*;
 
 /// Used to construct a [Pipeline].
 ///
@@ -263,7 +165,7 @@ impl PipelineBuilder {
         F: Fn(I) -> Fut + Send + Sync + 'static,
         Fut: Future<Output = ()> + Send + 'static,
     {
-        self.with_branching_stage(pipe, vec![], move |value| {
+        self.with_branching_stage(pipe, Vec::<String>::new(), move |value| {
             let task_fut = task(value);
             async move {
                 task_fut.await;
@@ -289,9 +191,13 @@ impl PipelineBuilder {
     ///
     /// This pipeline builder.
     ///
-    pub fn with_stage<S, I, O, F, Fut>(self, input_pipe: S, output_pipe: S, task: F) -> Self
+    pub fn with_stage<I, O, F, Fut>(
+        self,
+        input_pipe: impl AsRef<str>,
+        output_pipe: impl AsRef<str>,
+        task: F,
+    ) -> Self
     where
-        S: AsRef<str>,
         I: Send + 'static,
         O: Send + 'static,
         F: Fn(I) -> Fut + Send + Sync + 'static,
@@ -333,14 +239,13 @@ impl PipelineBuilder {
     ///
     /// This pipeline builder.
     ///
-    pub fn with_branching_stage<S, I, F, Fut>(
+    pub fn with_branching_stage<I, F, Fut>(
         mut self,
-        input_pipe: S,
-        output_pipes: Vec<S>,
+        input_pipe: impl AsRef<str>,
+        output_pipes: Vec<impl AsRef<str>>,
         task: F,
     ) -> Self
     where
-        S: AsRef<str>,
         I: Send + 'static,
         F: Fn(I) -> Fut + Send + Sync + 'static,
         Fut: Future<Output = Option<Vec<Option<BoxedAnySend>>>> + Send + 'static,
@@ -389,7 +294,8 @@ impl PipelineBuilder {
     /// use std::sync::atomic::{AtomicI32, Ordering};
     /// use async_pipes::Pipeline;
     ///
-    /// tokio_test::block_on(async {
+    /// #[tokio::main]
+    /// async fn main() {
     ///     let sum = Arc::new(AtomicI32::new(0));
     ///     let task_sum = sum.clone();
     ///
@@ -408,7 +314,7 @@ impl PipelineBuilder {
     ///         .await;
     ///
     ///     assert_eq!(sum.load(Ordering::Acquire), 15);
-    /// });
+    /// }
     /// ```
     ///
     /// # Returns
@@ -424,7 +330,6 @@ impl PipelineBuilder {
         It: IntoIterator + Send + 'static,
         It::Item: Send,
     {
-        // self.with_branching_flattener::<It>(from_pipe, vec![to_pipe])
         let input_pipe = from_pipe.as_ref().to_string();
         let output_pipe = to_pipe.as_ref().to_string();
         let err_pipe = input_pipe.clone();
@@ -443,6 +348,49 @@ impl PipelineBuilder {
             },
         }));
         self
+    }
+
+    /// A utility function for improving readability when building pipelines.
+    ///
+    /// This makes splitting up task definitions into functions easier by allowing the caller
+    /// to pass in a function that takes `self` and returns `self`, effectively providing
+    /// continuity to a call chain.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::sync::Arc;
+    /// use std::sync::atomic::{AtomicI32, Ordering};
+    /// use async_pipes::Pipeline;
+    /// use async_pipes::PipelineBuilder;
+    ///
+    /// #[tokio::main]
+    /// async fn main() {
+    ///     let sum = Arc::new(AtomicI32::new(0));
+    ///
+    ///     Pipeline::builder()
+    ///         .with_inputs("Numbers", vec![1, 2, 3, 4, 5])
+    ///         .also(build_consumer(sum.clone()))
+    ///         .build()
+    ///         .expect("failed to build pipeline")
+    ///         .wait()
+    ///         .await;
+    ///
+    ///     assert_eq!(sum.load(Ordering::Acquire), 15);
+    /// }
+    ///
+    /// fn build_consumer(sum: Arc<AtomicI32>) -> impl FnOnce(PipelineBuilder) -> PipelineBuilder {
+    ///     |builder| builder.with_consumer("Numbers", move |value: i32| {
+    ///         let sum = sum.clone();
+    ///         async move {
+    ///             sum.fetch_add(value, Ordering::SeqCst);
+    ///         }
+    ///     })
+    /// }
+    /// ```
+    ///
+    pub fn also(self, handler: impl FnOnce(PipelineBuilder) -> PipelineBuilder) -> Self {
+        handler(self)
     }
 
     /// When the pipeline is ready to be built, this is called and will return a pipeline if
@@ -689,370 +637,5 @@ impl PipelineBuilder {
         }
 
         tasks.shutdown().await;
-    }
-}
-
-/// A pipeline provides the infrastructure for managing a set of workers that run user-defined
-/// "tasks" on data going through the pipes.
-///
-/// # Examples
-///
-/// Creating a single producer and a single consumer.
-/// ```
-/// use std::sync::Arc;
-/// use std::sync::atomic::{AtomicUsize, Ordering};
-/// use std::sync::atomic::Ordering::{Acquire, SeqCst};
-/// use tokio::sync::Mutex;
-/// use async_pipes::Pipeline;
-///
-/// tokio_test::block_on(async {
-///     let count = Arc::new(Mutex::new(0usize));
-///
-///     let sum = Arc::new(AtomicUsize::new(0));
-///     let task_sum = sum.clone();
-///
-///     Pipeline::builder()
-///         // Produce values 1 through 10
-///         .with_producer("data", move || {
-///             let count = count.clone();
-///             async move {
-///                 let mut count = count.lock().await;
-///                 if *count < 10 {
-///                     *count += 1;
-///                     Some(*count)
-///                 } else {
-///                     None
-///                 }
-///             }
-///         })
-///         .with_consumer("data", move |value: usize| {
-///             let sum = task_sum.clone();
-///             async move {
-///                 sum.fetch_add(value, SeqCst);
-///             }
-///         })
-///         .build()
-///         .expect("failed to build pipeline")
-///         .wait()
-///         .await;
-///
-///     assert_eq!(sum.load(Acquire), 55);
-/// });
-/// ```
-///
-/// Creating a branching producer and two consumers for each branch.
-/// ```
-/// use std::sync::Arc;
-/// use std::sync::atomic::{AtomicUsize, Ordering};
-/// use std::sync::atomic::Ordering::Acquire;
-/// use tokio::sync::Mutex;
-/// use async_pipes::{BoxedAnySend, Pipeline};
-///
-/// tokio_test::block_on(async {
-///     let count = Arc::new(Mutex::new(0usize));
-///
-///     let odds_sum = Arc::new(AtomicUsize::new(0));
-///     let task_odds_sum = odds_sum.clone();
-///
-///     let evens_sum = Arc::new(AtomicUsize::new(0));
-///     let task_evens_sum = evens_sum.clone();
-///
-///     Pipeline::builder()
-///         .with_branching_producer(vec!["evens", "odds"], move || {
-///             let c = count.clone();
-///             async move {
-///                 let mut c = c.lock().await;
-///                 if *c >= 10 {
-///                     return None;
-///                 }
-///                 *c += 1;
-///
-///                 let output: Vec<Option<BoxedAnySend>> = if *c % 2 == 0 {
-///                     vec![Some(Box::new(*c)), None]
-///                 } else {
-///                     vec![None, Some(Box::new(*c))]
-///                 };
-///
-///                 Some(output)
-///             }
-///         })
-///         .with_consumer("odds", move |n: usize| {
-///             let odds_sum = task_odds_sum.clone();
-///             async move {
-///                 odds_sum.fetch_add(n, Ordering::SeqCst);
-///             }
-///         })
-///         .with_consumer("evens", move |n: usize| {
-///             let evens_sum = task_evens_sum.clone();
-///             async move {
-///                 evens_sum.fetch_add(n, Ordering::SeqCst);
-///             }
-///         })
-///         .build()
-///         .expect("failed to build pipeline!")
-///         .wait()
-///         .await;
-///
-///     assert_eq!(odds_sum.load(Acquire), 25);
-///     assert_eq!(evens_sum.load(Acquire), 30);
-/// });
-/// ```
-#[derive(Debug)]
-pub struct Pipeline {
-    synchronizer: Arc<Synchronizer>,
-    producers: JoinSet<()>,
-    workers: JoinSet<()>,
-    signal_txs: Vec<Sender<StageWorkerSignal>>,
-}
-
-impl Pipeline {
-    /// Create a new pipeline builder.
-    pub fn builder() -> PipelineBuilder {
-        PipelineBuilder::default()
-    }
-
-    /// Wait for the pipeline to complete.
-    ///
-    /// Once the pipeline is complete, a termination signal is sent to to all the workers.
-    ///
-    /// A pipeline progresses to completion by doing the following:
-    ///   1. Wait for all "producers" to complete while also progressing stage workers
-    ///   2. Wait for either all the stage workers to complete, or wait for the internal
-    ///      synchronizer to notify of completion (i.e. there's no more data flowing through the
-    ///      pipeline)
-    ///
-    /// Step 1 implies that if the producers never finish, the pipeline will run forever. See
-    /// [PipelineBuilder::with_producer] for more info.
-    pub async fn wait(mut self) {
-        let workers_to_progress = Arc::new(Mutex::new(self.workers));
-        let workers_to_finish = workers_to_progress.clone();
-
-        let wait_for_producers = async {
-            while let Some(result) = self.producers.join_next().await {
-                check_join_result(result);
-            }
-        };
-        let wait_for_workers = |workers: Arc<Mutex<JoinSet<()>>>| async move {
-            while let Some(result) = workers.lock().await.join_next().await {
-                check_join_result(result);
-            }
-        };
-        let check_sync_completed = async move {
-            while !self.synchronizer.completed() {
-                yield_now().await
-            }
-
-            for tx in self.signal_txs {
-                tx.send(StageWorkerSignal::Terminate)
-                    .await
-                    .expect("failed to send done signal")
-            }
-        };
-
-        // Effectively, make progress until all producers are done.
-        // We do this to prevent the synchronizer from causing the pipeline to shutdown too early.
-        select! {
-            _ = wait_for_producers => {},
-            _ = wait_for_workers(workers_to_progress),
-                if !workers_to_progress.lock().await.is_empty() => {},
-        }
-
-        // If either the synchronizer determines we're done, or all workers completed, we're done
-        select! {
-            _ = wait_for_workers(workers_to_finish) => {},
-            _ = check_sync_completed => {},
-        }
-    }
-}
-
-fn find_reader(
-    name: &str,
-    pipes: &mut HashMap<String, Pipe<BoxedAnySend>>,
-) -> Result<PipeReader<BoxedAnySend>, String> {
-    Ok(pipes
-        .get_mut(name)
-        .unwrap_or_else(|| panic!("no pipe with name '{}' found", name))
-        .reader
-        .take()
-        .ok_or("reader was already used")?)
-}
-
-fn find_writer(
-    name: &str,
-    pipes: &HashMap<String, Pipe<BoxedAnySend>>,
-) -> Result<PipeWriter<BoxedAnySend>, String> {
-    Ok(pipes
-        .get(name)
-        .ok_or(&format!("pipeline has open-ended pipe: '{}'", name))?
-        .writer
-        .clone())
-}
-
-fn find_writers(
-    names: &[String],
-    pipes: &HashMap<String, Pipe<BoxedAnySend>>,
-) -> Result<Vec<PipeWriter<BoxedAnySend>>, String> {
-    let mut writers = Vec::new();
-    for name in names {
-        writers.push(find_writer(name, pipes)?);
-    }
-    Ok(writers)
-}
-
-async fn write_results<O>(writers: &[PipeWriter<O>], results: Vec<Option<O>>) {
-    if results.len() != writers.len() {
-        panic!("len(results) != len(writers)");
-    }
-
-    for (result, writer) in results.into_iter().zip(writers) {
-        if let Some(result) = result {
-            writer.write(result).await;
-        }
-    }
-}
-
-fn downcast_from_pipe<T: 'static>(value: BoxedAnySend, pipe_name: &str) -> Box<T> {
-    value.downcast::<T>().unwrap_or_else(|_| {
-        panic!(
-            "failed to downcast input value to {} from pipe '{}'",
-            type_name::<T>(),
-            pipe_name,
-        )
-    })
-}
-
-fn check_join_result<T>(result: Result<T, JoinError>) {
-    if let Err(e) = result {
-        if e.is_panic() {
-            panic::resume_unwind(e.into_panic())
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::atomic::Ordering::{Acquire, Release, SeqCst};
-    use std::sync::atomic::{AtomicBool, AtomicUsize};
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    use tokio::select;
-    use tokio::sync::mpsc::channel;
-
-    use crate::{Pipeline, StageWorkerSignal};
-
-    #[tokio::test]
-    async fn test_stage_producer() {
-        let written = Arc::new(AtomicBool::new(false));
-        let task_written = written.clone();
-
-        Pipeline::builder()
-            .with_producer("pipe", move || {
-                let written = task_written.clone();
-                async move {
-                    if !written.load(Acquire) {
-                        written.store(true, Release);
-                        Some("hello!".to_string())
-                    } else {
-                        None
-                    }
-                }
-            })
-            .with_consumer("pipe", |value: String| async move {
-                assert_eq!(value, "hello!".to_string());
-            })
-            .build()
-            .unwrap()
-            .wait()
-            .await;
-
-        assert!(written.load(Acquire), "value was not handled by worker!");
-    }
-
-    #[tokio::test]
-    async fn test_stage_single_output() {
-        let written = Arc::new(AtomicBool::new(false));
-        let task_written = written.clone();
-
-        Pipeline::builder()
-            .with_inputs("first", vec![1usize])
-            .with_stage(
-                "first",
-                "second",
-                |value: usize| async move { Some(value + 1) },
-            )
-            .with_consumer("second", move |value: usize| {
-                let written = task_written.clone();
-                async move {
-                    assert_eq!(value, 2);
-                    written.store(true, Release);
-                }
-            })
-            .build()
-            .unwrap()
-            .wait()
-            .await;
-
-        assert!(written.load(Acquire), "value was not handled by worker!");
-    }
-
-    #[tokio::test]
-    async fn test_stage_flattener() {
-        let sum = Arc::new(AtomicUsize::new(0));
-        let task_sum = sum.clone();
-
-        Pipeline::builder()
-            .with_inputs("first", vec![vec![1usize, 2usize, 3usize]])
-            .with_flattener::<Vec<usize>>("first", "second")
-            .with_consumer("second", move |value: usize| {
-                let sum = task_sum.clone();
-                async move {
-                    sum.fetch_add(value, SeqCst);
-                }
-            })
-            .build()
-            .unwrap()
-            .wait()
-            .await;
-
-        assert_eq!(sum.load(Acquire), 6);
-    }
-
-    #[tokio::test]
-    #[should_panic]
-    async fn test_stage_propagates_task_panic() {
-        Pipeline::builder()
-            .with_inputs("pipe", vec![()])
-            .with_consumer("pipe", |_: ()| async move { panic!("AHH!") })
-            .build()
-            .unwrap()
-            .wait()
-            .await;
-    }
-
-    #[tokio::test]
-    async fn test_stage_receives_signal_terminate() {
-        let (tx, mut rx) = channel(1);
-
-        let pipeline = Pipeline::builder()
-            .with_inputs("pipe", vec![()])
-            .with_consumer("pipe", move |_: ()| {
-                let tx = tx.clone();
-                async move {
-                    tx.send(()).await.unwrap();
-                    tokio::time::sleep(Duration::from_secs(5)).await;
-                    panic!("worker did not terminate!");
-                }
-            })
-            .build()
-            .unwrap();
-
-        let signaller = pipeline.signal_txs.first().unwrap().clone();
-        select! {
-            _ = pipeline.wait() => {},
-            _ = rx.recv() => {
-                signaller.send(StageWorkerSignal::Terminate).await.unwrap();
-            }
-        }
     }
 }
